# openapi.yaml
openapi: 3.1.0
info:
  title: Datamart API
  version: "1.4"
servers:
  - url: https://chat-gdatabot.onrender.com
security:
  - BearerAuth: []
paths:
  /peoplecounting:
    get:
      operationId: getPeopleCounting
      tags:
        - Shopping
      summary: Consulta o fluxo sumarizado de acesso de pessoas ao shopping
      parameters:
        - name: start_date
          in: query
          required: false
          description: Data inicial no formato AAAA-MM-DD
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: false
          description: Data final no formato AAAA-MM-DD
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Lista de registros
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PeopleCountingItem"

  /parkingdata:
    get:
      operationId: getParkingData
      tags:
        - Fluxos Mall
      summary: Consulta o fluxo sumarizado de acesso de veículos ao shopping
      security:
        - TokenHeader: []
      parameters:
        - name: start_date
          in: query
          required: false
          description: Data inicial no formato AAAA-MM-DD
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: false
          description: Data final no formato AAAA-MM-DD
          schema:
            type: string
            format: date
        - name: limit
          in: query
          required: false
          description: Máximo de registros a retornar
          schema:
            type: integer
            default: 1000
        - name: last_date
          in: query
          required: false
          description: Data da última entrada
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Lista de registros
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParkingDataItem"

  /hotspotaccess:
    get:
      operationId: getHotspotAccess
      tags:
        - Shopping
      summary: Consulta dados sumarizados de acesso ao hotspot de wi-fi
      security:
        - TokenHeader: []
      parameters:
        - name: start_date
          in: query
          required: false
          description: Data inicial no formato AAAA-MM:Semana X
          schema:
            type: string
        - name: end_date
          in: query
          required: false
          description: Data final no formato AAAA-MM:Semana N
          schema:
            type: string
        - name: last_date
          in: query
          required: false
          description: Data da última entrada do lote anterior (formato AAAA-MM:Semana N)
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Máximo de registros a retornar
          schema:
            type: integer
            default: 1000
      responses:
        "200":
          description: Lista de registros
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HotspotAccessItem"

  /vendas:
    get:
      operationId: getVendas
      tags:
        - Financeiro
      summary: Consultar dados de venda
      security:
        - BearerAuth: []
      parameters:
        - name: start_date
          in: query
          required: false
          description: Data inicial no formato AAAA-MM-DD
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: false
          description: Data final no formato AAAA-MM-DD
          schema:
            type: string
            format: date
        - name: last_date
          in: query
          required: false
          description: Data da última entrada do lote anterior (para paginação)
          schema:
            type: string
            format: date
        - name: limit
          in: query
          required: false
          description: Máximo de registros a retornar
          schema:
            type: integer
            default: 1000
      responses:
        "200":
          description: Lista de registros
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VendasItem"

  /receitasienge:
    post:
      operationId: registrarReceita
      tags:
        - Bairro Harmonia
      summary: Insere dados na tabela receitasienge
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceitaSiengeInput"
      responses:
        "200":
          description: Registro inserido com sucesso na tabela receitasienge
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string


components:
  securitySchemes:
    BearerAuth:
      type: http
      in: bearer
      bearerFormat: JWT

  schemas:

    PeopleCountingItem:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        count:
          type: integer
        location:
          type: string

    ParkingDataItem:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        plate:
          type: string
        zone:
          type: string

    HotspotAccessItem:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        mac_address:
          type: string
        device_type:
          type: string

    VendasItem:
      type: object
      properties:
        loja:
          type: string
        luc:
          type: string
        dt_referencia:
          type: string
          format: date
        vl_vendido:
          type: number
        categoria:
          type: string
        segmento:
          type: string
        classificacao:
          type: string
        semana:
          type: string

    ReceitaSiengeInput:
      type: object
      properties:
        dt_vencimento:
          type: string
          format: date
        cliente:
          type: string
        documento:
          type: string
        titulo:
          type: integer
        parcela:
          type: string
        tc:
          type: string
        unidade:
          type: string
        vl_original:
          type: number
          format: float
        dt_calculo:
          type: string
          format: date
        sa_atual:
          type: number
          format: float
        acrescimo:
          type: number
          format: float
        desconto:
          type: number
          format: float
        vl_total:
          type: number
          format: float